requirements.txt
amqp==2.4.2
appnope==0.1.0
backcall==0.1.0
billiard==3.5.0.5
boto3==1.9.125
botocore==1.12.125
cached-property==1.5.1
celery==4.2.1
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
django==2.1.9
django-cors-headers==2.5.2
django-hosts==3.0
django-redis==4.10.0
django-storages==1.7.1
django-url-filter==0.3.12
djangorestframework==3.9.2
docutils==0.14
enum-compat==0.0.2
gunicorn==19.9.0
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
jmespath==0.9.4
kombu==4.4.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.9
raven==6.10.0
redis==3.2.1
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7